requirements.txt
pyrebase4
pandas
streamlit
requests
streamlit-option-menu
firebase-admin
st-pages
google-cloud-firestore
google-generativeai==0.1.0rc1